PIC-3159 Add test config for feign client
