+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/config/FeignConfig.kt
@@ -1,7 +1,9 @@
 package uk.gov.justice.digital.hmpps.personrecord.config
 
+import com.fasterxml.jackson.databind.JsonNode
 import feign.RequestInterceptor
 import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Profile
 import org.springframework.http.HttpHeaders
 import org.springframework.security.authentication.AnonymousAuthenticationToken
 import org.springframework.security.core.authority.AuthorityUtils
@@ -9,6 +11,7 @@ import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.oauth2.client.OAuth2AuthorizeRequest
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClientManager
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException
+import org.springframework.web.client.RestTemplate
 
 abstract class FeignConfig(
   private val authorizedClientManager: OAuth2AuthorizedClientManager,
@@ -17,8 +20,20 @@ abstract class FeignConfig(
   abstract fun registrationId(): String
 
   @Bean
+  @Profile("!test")
   open fun requestInterceptor() = RequestInterceptor { template ->
     template.header(HttpHeaders.AUTHORIZATION, "Bearer ${getAccessToken()}")
+  }
+
+  @Bean
+  @Profile("test")
+  open fun requestInterceptorForTest() = RequestInterceptor { template ->
+    template.header(HttpHeaders.AUTHORIZATION, "Bearer ${getLocalAccessToken()}")
+  }
+  private fun getLocalAccessToken(): String {
+    val authResponse = RestTemplate()
+      .postForObject("http://localhost:8090/auth/oauth/token", null, JsonNode::class.java)!!
+    return authResponse["access_token"].asText()
   }
 
   private fun getAccessToken(): String {
