+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/config/SecurityConfiguration.kt
@@ -4,6 +4,11 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
+import org.springframework.security.oauth2.client.OAuth2AuthorizedClientManager
+import org.springframework.security.oauth2.client.OAuth2AuthorizedClientProviderBuilder
+import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository
+import org.springframework.security.oauth2.client.web.DefaultOAuth2AuthorizedClientManager
+import org.springframework.security.oauth2.client.web.OAuth2AuthorizedClientRepository
 import org.springframework.security.web.SecurityFilterChain
 
 @Configuration
@@ -35,5 +40,24 @@ class SecurityConfiguration {
       }
 
     return http.build()
+  }
+
+  @Bean
+  fun authorizedClientManager(
+    clientRegistrationRepository: ClientRegistrationRepository?,
+    authorizedClientRepository: OAuth2AuthorizedClientRepository?,
+  ): OAuth2AuthorizedClientManager {
+    val authorizedClientProvider = OAuth2AuthorizedClientProviderBuilder.builder()
+      .clientCredentials()
+      .build()
+
+    val authorizedClientManager = DefaultOAuth2AuthorizedClientManager(
+      clientRegistrationRepository,
+      authorizedClientRepository,
+    )
+
+    authorizedClientManager.setAuthorizedClientProvider(authorizedClientProvider)
+
+    return authorizedClientManager
   }
 }
