+++ src/test/resources/__files/oauth/hmpps-auth-token-body.json
@@ -1,0 +1,11 @@
+{
+  "access_token": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJwcm9iYXRpb24taW50ZWdyYXRpb24tZGV2IiwiZ3JhbnRfdHlwZSI6ImNsaWVudF9jcmVkZW50aWFscyIsInVzZXJfbmFtZSI6InByb2JhdGlvbi1pbnRlZ3JhdGlvbi1kZXYiLCJzY29wZSI6WyJyZWFkIiwid3JpdGUiXSwiYXV0aF9zb3VyY2UiOiJub25lIiwiaXNzIjoiaHR0cHM6Ly9zaWduLWluLWRldi5obXBwcy5zZXJ2aWNlLmp1c3RpY2UuZ292LnVrL2F1dGgvaXNzdWVyIiwiZXhwIjo5OTk5OTk5OTk5LCJhdXRob3JpdGllcyI6WyJST0xFX0FQUFJPVkVEX1BSRU1JU0VTX0FTU0VTU01FTlRTIl0sImp0aSI6IjI1RHVSbjEtaHlIWmV3TGNkSkp4d1ZMMDNLVSIsImNsaWVudF9pZCI6InByb2JhdGlvbi1pbnRlZ3JhdGlvbi1kZXYiLCJpYXQiOjE2NjM3NTczMTF9.teMLHccVosDs8-v2nhsfiYg66TcL44oKtmNlVAybQncrWRG66LpSkZyf7ashCEuYrSd_Thqh1JRQsrLtWdGAA384tAw-kRQOG-67HYsXWXUn3O8VizxSJM-Ng75IeHBX5c3eYlhYYhxFYoL4H6uMuQog4BaBaHD63ZgovPhG3Pw",
+  "token_type": "bearer",
+  "expires_in": 9999999999,
+  "scope": "read write",
+  "sub": "hmpps-person-record-dev",
+  "auth_source": "none",
+  "jti": "fN29JHJy1N7gcYvqe-8B_k5T0mA",
+  "iss": "https://sign-in-dev.hmpps.service.justice.gov.uk/auth/issuer"
+}
+
