+++ src/test/resources/mappings/oauth/auth.json
@@ -1,0 +1,17 @@
+{
+  "mappings": [
+    {
+      "request": {
+        "method": "POST",
+        "urlPath": "/auth/oauth/token"
+      },
+      "response": {
+        "headers": {
+          "Content-Type": "application/json"
+        },
+        "status": 200,
+        "bodyFileName": "oauth/hmpps-auth-token-body.json"
+      }
+    }
+  ]
+}
